requirements.txt
pythermalcomfort
python-dotenv
Alembic
psycopg2-binary
SQLAlchemy
pytz
requests
i18nice[YAML]
pydantic